ランキングの実装

+++ project/Source/Gunsetting.h
@@ -24,5 +24,5 @@ private:
 	float maxammoBuf;
 	float bomAtkBuf;
 	float bomRangeBuf;
-	int   bomAmmoBuf;
+	float bomAmmoBuf;
 };
